Moved base files into main package + Renamed EventManager into ListenersManager
