+++ src/main/java/fr/wydavix/WydavixTools/Utils.java
