+++ src/main/java/fr/wydavix/WydavixTools/WarningsEvent.java
